requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
argcomplete==1.10.3
asn1crypto==1.5.1
bcrypt==4.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
compressed-rtf==1.0.7
cryptography==45.0.3
cssselect2==0.8.0
defusedxml==0.7.1
docx2txt==0.9
ebcdic==1.1.1
extract-msg==0.28.7
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
fonttools==4.59.0
fpdf==1.7.2
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
hf-xet==1.1.5
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.34.3
ibm-cos-sdk==2.14.2
ibm-cos-sdk-core==2.14.2
ibm-cos-sdk-s3transfer==2.14.2
ibm_watson_machine_learning==1.0.368
idna==3.10
IMAPClient==2.1.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
lomond==0.3.3
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
networkx==3.5
numpy==1.26.4
olefile==0.47
oscrypto==1.3.0
packaging==25.0
pandas==2.1.4
pdfminer.six==20191110
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
pyHanko==0.28.0
pyhanko-certvalidator==0.27.0
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
regex==2025.7.31
reportlab==4.4.1
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
textract==1.6.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.4.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
xlrd==1.2.0
xlsxwriter==3.2.5
zipp==3.23.0
zopfli==0.2.3.post1
